kanban/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Authlib==1.6.4
bandit==1.8.6
bcrypt==5.0.0
billiard==4.2.2
black==25.9.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.7
cryptography==46.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
dparse==0.6.4
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
filelock==3.19.1
flake8==7.3.0
frozenlist==1.7.0
groq==0.32.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langsmith==0.4.31
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.4
mypy==1.18.2
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.3.3
ollama==0.6.0
openai==2.6.0
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.3
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safety==3.6.2
safety-schemas==0.0.16
sendgrid==6.12.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
stevedore==5.5.0
stripe==12.5.1
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.25.0
